requirements.txt
streamlit==0.79.0
pandas==1.2.3
Pillow==8.2.0
